+++ protected/controllers/ThreadController.php
@@ -32,12 +32,8 @@ class ThreadController extends Controller
 				'users'=>array('@'),
 			),
 			array('allow', // allow authenticated user to perform 'create' and 'update' actions
-				'actions'=>array('create','update'),
+				'actions'=>array('create','update','admin','delete'),
 				'users'=>array('@'),
-			),
-			array('allow', // allow admin user to perform 'admin' and 'delete' actions
-				'actions'=>array('admin','delete'),
-				'users'=>array('admin'),
 			),
 			array('deny',  // deny all users
 				'users'=>array('*'),
@@ -69,6 +65,9 @@ class ThreadController extends Controller
 
 		if(isset($_POST['Thread']))
 		{
+			$model->user_id=Yii::app()->user->id;
+			$model->created_at=time();
+			$model->updated_at=time();
 			$model->attributes=$_POST['Thread'];
 			if($model->save())
 				$this->redirect(array('view','id'=>$model->id));
@@ -86,6 +85,10 @@ class ThreadController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
+		if(!Yii::app()->user->checkAccess('admin'))
+		{
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 		$model=$this->loadModel($id);
 
 		// Uncomment the following line if AJAX validation is needed
@@ -110,6 +113,10 @@ class ThreadController extends Controller
 	 */
 	public function actionDelete($id)
 	{
+		if(!Yii::app()->user->checkAccess('admin'))
+		{
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 		$this->loadModel($id)->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
@@ -133,6 +140,10 @@ class ThreadController extends Controller
 	 */
 	public function actionAdmin()
 	{
+		if(!Yii::app()->user->checkAccess('admin'))
+		{
+			throw new CHttpException(403,'You are not authorized to perform this action.');
+		}
 		$model=new Thread('search');
 		$model->unsetAttributes();  // clear any default values
 		if(isset($_GET['Thread']))
